requirements.txt
alembic==1.7.7
click==8.1.2
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-gitlab==3.3.0
requests==2.27.1
SQLAlchemy==1.4.35
Werkzeug==2.1.1
